pyproject.toml
[tool.poetry]
name = "fightingice-ai"
version = "0.1.0"
description = ""
authors = ["Robin Gieseke <robin.gieseke@me.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
river = "^0.19.0"
numpy = "^1.26.1"
pandas = "^2.1.2"
scikit-learn = "^1.3.2"
pytest = "^7.4.3"
pyftg = {git = "https://github.com/TheDarkchip/pyftg.git"}
stable-baselines3 = "^2.1.0"
protobuf = "^4.25.0"
tensorflow = ">=2.3.0"
autokeras = "^1.1.0"
visualkeras = "^0.0.2"
pillow = "9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
